remove track child views
