+++ packages/author-profile/author-profile-content.web.js
@@ -1,8 +1,6 @@
 import React from "react";
 import { StyleSheet, View } from "react-native";
 import PropTypes from "prop-types";
-import _round from "lodash.round";
-import { addTracking } from "@times-components/tracking";
 import AuthorProfileHeader, {
   AuthorProfileHeaderWithTracking
 } from "./author-profile-header";
@@ -46,18 +44,8 @@ const makeAuthorProfileContent = Header => {
   return AuthorProfile;
 };
 
-export const AuthorProfileContentWithTracking = addTracking(
-  makeAuthorProfileContent(AuthorProfileHeaderWithTracking),
-  {
-    trackChildViews: {
-      id: "id",
-      attrs: {
-        index: props => props.index,
-        progress: props => _round((props.index + 1) / props.total * 100, 2)
-      },
-      listPath: "articles.list"
-    }
-  }
+export const AuthorProfileContentWithTracking = makeAuthorProfileContent(
+  AuthorProfileHeaderWithTracking
 );
 
 export default makeAuthorProfileContent(AuthorProfileHeader);
