+++ packages/author-profile/author-profile-item.js
@@ -12,7 +12,6 @@ const styles = StyleSheet.create({
 
 const AuthorProfileItem = item => {
   const props = {
-    id: item.id,
     date: item.publishedTime,
     headline: item.title,
     image: {
