+++ packages/author-profile/package.json
@@ -72,7 +72,6 @@
     "jest-cli": "20.0.4",
     "lodash.get": "4.4.2",
     "lodash.merge": "4.6.0",
-    "lodash.round": "4.0.4",
     "prop-types": "15.5.10",
     "react-native-web": "0.0.119",
     "react-style-proptype": "3.0.0"
