+++ packages/card/card.js
@@ -51,7 +51,6 @@ class CardComponent extends React.Component {
   render() {
     const { isHorizontal } = this.state;
     const {
-      id,
       date,
       headline,
       image,
@@ -76,7 +75,6 @@ class CardComponent extends React.Component {
 
     return (
       <View
-        id={id}
         onLayout={this.handleLayout}
         style={[
           styles.container,
