+++ packages/tracking/tracking-context.js
@@ -4,8 +4,7 @@ import _get from "lodash.get";
 import hoistNonReactStatic from "hoist-non-react-statics";
 import resolveAttrs from "./resolve-attrs";
 import getDisplayName from "./get-display-name";
-import { trackingContextTypes } from "./base-tracking";
-import { addTracking } from "./tracking";
+import { addTracking, trackingContextTypes } from "./tracking";
 
 const addTrackingContext = (WrappedComponent, { attrs = {} } = {}) => {
   class WithTrackingContext extends Component {
